+++ jobs/impute_ind_cqc_ascwds_and_pir.py
@@ -8,7 +8,7 @@ from utils.column_names.ind_cqc_pipeline_columns import (
     IndCqcColumns as IndCQC,
     PartitionKeys as Keys,
 )
-from utils.estimate_filled_posts.models.primary_service_rolling_rate_of_change import (
+from utils.estimate_filled_posts.models.primary_service_rate_of_change import (
     model_primary_service_rate_of_change,
 )
 from utils.estimate_filled_posts.models.imputation_with_extrapolation_and_interpolation import (
